fix: migration change as per sheet which given by kalpeshbhai

+++ database/migrations/2023_03_05_115658_create_college_timetable_table.php
@@ -14,16 +14,16 @@ return new class extends Migration {
     {
         Schema::create('college_timetable', function (Blueprint $table) {
             $table->comment('');
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->integer('sub_institute_id');
             $table->integer('syear');
-            $table->unsignedInteger('academic_section_id')->index('timetable_academic_section_id_foreign');
-            $table->unsignedInteger('standard_id')->index('timetable_standard_id_foreign');
-            $table->unsignedInteger('division_id')->index('timetable_division_id_foreign');
-            $table->unsignedInteger('batch_id')->nullable();
-            $table->unsignedInteger('period_id')->index('timetable_period_id_foreign');
-            $table->unsignedInteger('subject_id')->index('timetable_subject_id_foreign');
-            $table->unsignedInteger('teacher_id')->index('timetable_teacher_id_foreign');
+            $table->bigInteger('academic_section_id')->index('timetable_academic_section_id_foreign');
+            $table->bigInteger('standard_id')->index('timetable_standard_id_foreign');
+            $table->bigInteger('division_id')->index('timetable_division_id_foreign');
+            $table->bigInteger('batch_id')->nullable();
+            $table->bigInteger('period_id')->index('timetable_period_id_foreign');
+            $table->bigInteger('subject_id')->index('timetable_subject_id_foreign');
+            $table->bigInteger('teacher_id')->index('timetable_teacher_id_foreign');
             $table->string('week_day', 255);
             $table->string('extended', 10)->nullable();
             $table->string('type', 50)->nullable();
